+++ changelog.md
@@ -2,6 +2,7 @@
 Wathe 1.3.1 - 1.21.1
 ------------------------------------------------------
 - Fixed map effect not initializing when being manually changed with the mapVariables command
+- Fixed time of day not properly changing with map effects other than the default Harpy Express ones
 
 ------------------------------------------------------
 Wathe 1.3 - 1.21.1